% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{distribution}
\alias{distribution}
\title{Gets the current XLM distribution stats.}
\usage{
distribution()
}
\value{
list
}
\description{
Uses an API call from the Stellar dashboard to get the distribution figures.
}
\note{
https://dashboard.stellar.org/
}
\examples{
xlm_stats = distribution()
}
