% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friendbot.R
\name{friendbot}
\alias{friendbot}
\title{Creates and fund a new account on the test network.}
\usage{
friendbot(account)
}
\arguments{
\item{account}{character - the public key of the account you wish to fund.}
}
\value{
list
}
\description{
Make a call to the Friendbot service on the test network and fund an account.
    Returns a JSON response.
}
\note{
https://www.stellar.org/developers/guides/get-started/create-account.html
}
\examples{
friendbot("GAKYVLEN4MFPQ2R5Y4BA5RSJ7U5R2HGOK3VJ5PEUTLOHKTCBGCJZSHOS")
}
