% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLedgerDetail.R
\name{getLedgerDetail}
\alias{getLedgerDetail}
\title{Returns a single ledger.}
\usage{
getLedgerDetail(sequence, domain = "public")
}
\arguments{
\item{sequence}{character - (required) a ledger id. Must be 1 or greater.}

\item{domain}{character - signalling to query the testnet or public network. Can be either "testnet" or "public".}
}
\value{
list
}
\description{
Return a single ledger by providing a sequence id.
}
\note{
https://www.stellar.org/developers/horizon/reference/endpoints/ledgers-single.html
}
\examples{
\dontrun{getLedgerDetail("1", "testnet")}
}
