% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ledger.R
\docType{data}
\name{Ledger}
\alias{Ledger}
\title{Ledger object.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Ledger
}
\description{
An R6 class representing a Ledger on the Stellar network.
    It is initialized from the getLedgerDetail() function.
}
\section{Fields}{

\describe{
\item{\code{hash}}{- a cryptographic signature of all of the data inside the ledger.}

\item{\code{sequence}}{- the current sequence number of the ledger}

\item{\code{transaction_count}}{- how many transactions in the ledger.}

\item{\code{closed_at}}{- when the transaction was confirmed.}

\item{\code{total_coins}}{- how many lumens existed on the network at the time the ledger was confirmed.}

\item{\code{response}}{- list containing the raw JSON response.}
}}

\note{
https://www.stellar.org/developers/guides/concepts/ledger.html
}
\examples{
genesis <- Ledger$new("1")
genesis$effects()
}
\keyword{datasets}
