% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listToDF.R
\name{listToDF}
\alias{listToDF}
\title{Convert parsed JSON responses to a tabular format.}
\usage{
listToDF(response)
}
\arguments{
\item{response}{list - the list object returned by the other xlm functions.}
}
\value{
one or more data.tables
}
\description{
Takes a successful response containing multiple records and converts it into a table format.
    It works by flattening semi-structured lists into a table where it can.
    For example, if there are multiple signatures to a transaction then it will warn the user before moving those values to a separate table.
}
