% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Orderbook.R
\docType{data}
\name{Orderbook}
\alias{Orderbook}
\title{Order book object.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Orderbook
}
\description{
An R6 class representing an orderbook on the Stellar network.
    It is initialized from the getOrderbook() function.
}
\section{Fields}{

\describe{
\item{\code{bids}}{- a cryptographic signature of all of the data inside the ledger.}

\item{\code{asks}}{- the current sequence number of the ledger}

\item{\code{base}}{- how many transactions in the ledger.}

\item{\code{counter}}{- when the transaction was confirmed.}

\item{\code{domain}}{- character - signalling to query the testnet or public network. Can be either "testnet" or "public".}

\item{\code{response}}{- list containing the raw JSON response.}
}}

\note{
https://www.stellar.org/developers/horizon/reference/resources/orderbook.html
}
\keyword{datasets}
