% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postTransaction.R
\name{postTransaction}
\alias{postTransaction}
\title{Post a transaction.}
\usage{
postTransaction(body, domain = "public")
}
\arguments{
\item{body}{character - base64 representation of transaction envelope XDR.}

\item{domain}{- character - signalling to query the testnet or public network. Can be either "testnet" or "public".}
}
\value{
list
}
\description{
Posts a new transaction to the Stellar Network.
}
\note{
https://www.stellar.org/developers/horizon/reference/endpoints/transactions-create.html
}
\examples{
\dontrun{postTransaction("aaaaaafds098as0d9")}

}
