% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/price.R
\name{price}
\alias{price}
\title{Gets the current XLM price.}
\usage{
price(currency = "USDT", live = FALSE)
}
\arguments{
\item{currency}{character - current pairings are USD, ETH, BTC and BNB. USD is worked out indirectly via BTC with two API calls.}

\item{live}{boolean - if true, a while loop will continuously get the latest price and print it to the screen.}
}
\value{
numeric
}
\description{
Make a call to the Binance API and get the current XLM market price.
}
\note{
Details of API can be found here: https://github.com/binance-exchange/binance-official-api-docs/blob/master/rest-api.md
}
\examples{
price("USD")
}
